+++ src/components/pages/account/signup/SignUpCode/index.module.scss
